requirements.txt
streamlit==1.31.0
google-generativeai==0.3.2
beautifulsoup4==4.12.3
pdfplumber==0.11.0
